fix import for new exercise and add the old exercises in app.js
