+++ JavaScript/exercise2_3_4/exercise2&3&4.js
@@ -1,103 +1,8 @@
-// DTO
-class Article {
-    constructor(title, author, date) {
-        this.title = title;
-        this.author = author;
-        this.date = date;
-    }
-
-    toHTML() {
-        const li = document.createElement("li");
-        li.classList.add("li-article");
-        const title = document.createElement("h1");
-        title.classList.add("title-in-decimal-list");
-        const authorAndDate = document.createElement("p");
-
-        const year = this.date.getFullYear();
-        const month = String(this.date.getMonth() + 1).padStart(2, '0');
-        const day = String(this.date.getDate()).padStart(2, '0');
-        const formatedDate = `${year}-${month}-${day}`;
-
-        title.textContent = this.title;
-        authorAndDate.textContent = this.author + ", " + formatedDate;
-
-        li.appendChild(title);
-        li.appendChild(authorAndDate);
-
-        return li;
-    }
-}
-
-
-class ListOfArticles {
-    constructor(listOfArticles = []) {
-        this.ol = document.createElement("ol");
-        this.listOfArticles = listOfArticles;
-        this.orderByAuthorAscend = true;
-        this.orderByDateAscend = true;
-    }
-
-    addElement(article) {
-        this.listOfArticles.push(article);
-
-        this.display();
-    }
-    display() {
-        this.ol.innerHTML = "";
-
-        for (let article of this.listOfArticles) {
-            const articleToHTML = article.toHTML();
-
-            articleToHTML.onclick = () => {
-                const userConfirm = confirm(
-                    articleToHTML.firstElementChild.textContent +
-                    " ------ Do you want to delete this entry?"
-                );
-
-                if (userConfirm === true) {
-                    articleToHTML.remove();
-                    const index = this.listOfArticles.indexOf(article);
-                    this.listOfArticles.splice(index, 1);
-                }
-            };
-
-            this.ol.appendChild(articleToHTML);
-        }
-
-        return this.ol;
-    }
-
-    sortByAuthor() {
-        if (this.orderByAuthorAscend) {
-            this.listOfArticles.sort((a, b) => a.author.localeCompare(b.author));
-            this.orderByAuthorAscend = false;
-        } else {
-            this.listOfArticles.sort((a, b) => b.author.localeCompare(a.author));
-            this.orderByAuthorAscend = true;
-        }
-
-        this.display();
-    }
-
-    sortByDate() {
-        if (this.orderByDateAscend) {
-            this.listOfArticles.sort((a, b) => a.date - b.date);
-            this.orderByDateAscend = false;
-        } else {
-            this.listOfArticles.sort((a, b) => b.date - a.date);
-            this.orderByDateAscend = true;
-        }
-
-        this.display();
-    }
-}
-
-
-
-
+import Article from './models/Article.js';
+import ListOfArticles from './models/ListOfArticles.js';
 
 // Entry point of exercise
-function exercise2_3() {
+export default function exercise2_3_4() {
     const result = document.createElement("div")
 
     // Display elements
